Modify Store dropdown
