+++ frontend/spec/helpers/store_helper_spec.rb
@@ -1,0 +1,21 @@
+require 'spec_helper'
+
+module Spree
+  describe StoreHelper, type: :helper do
+    describe '#stores' do
+
+    end
+
+    describe '#store_country_iso' do
+
+    end
+
+    describe '#store_currency_symbol' do
+
+    end
+
+    describe '#store_locale_name' do
+
+    end
+  end
+end
